pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Ali May", email = "2628839532@qq.com"},
]
dependencies = ["fastapi[all]>=0.116.1", "sqlalchemy>=2.0.41", "psycopg2-binary>=2.9.10", "pydantic>=2.11.7", "pydantic-settings>=2.10.1", "PyJWT>=2.10.1"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
